+++ qa/rpc-tests/reorg_limit.py
@@ -18,7 +18,10 @@ from test_framework.util import (
 import tempfile
 from time import sleep
 
-def check_stopped(i, timeout=10):
+# Must match MAX_REORG_LENGTH in src/main.h.
+MAX_REORG_LENGTH = 1000
+
+def check_stopped(i, timeout=240):
     stopped = False
     for x in range(1, timeout):
         ret = check_node(i)
@@ -49,36 +52,37 @@ class ReorgLimitTest(BitcoinTestFramework):
         self.split_network()
 
         print("Test the maximum-allowed reorg:")
-        print("Mine 99 blocks on Node 0")
-        self.nodes[0].generate(99)
-        assert(self.nodes[0].getblockcount() == 299)
+        print("Mine %d blocks on Node 0" % MAX_REORG_LENGTH)
+        self.nodes[0].generate(MAX_REORG_LENGTH)
+        assert(self.nodes[0].getblockcount() == 200 + MAX_REORG_LENGTH)
         assert(self.nodes[2].getblockcount() == 200)
 
-        print("Mine competing 100 blocks on Node 2")
-        self.nodes[2].generate(100)
-        assert(self.nodes[0].getblockcount() == 299)
-        assert(self.nodes[2].getblockcount() == 300)
+        print("Mine competing %d blocks on Node 2" % (MAX_REORG_LENGTH + 1))
+        self.nodes[2].generate(MAX_REORG_LENGTH + 1)
+        assert(self.nodes[0].getblockcount() == 200 + MAX_REORG_LENGTH)
+        assert(self.nodes[2].getblockcount() == 201 + MAX_REORG_LENGTH)
 
         print("Connect nodes to force a reorg")
         connect_nodes_bi(self.nodes, 0, 2)
         self.is_network_split = False
-        sync_blocks(self.nodes)
+        # A reorg of ~MAX_REORG_LENGTH blocks takes well over the default timeout.
+        sync_blocks(self.nodes, timeout=600)
 
         print("Check Node 0 is still running and on the correct chain")
-        assert(self.nodes[0].getblockcount() == 300)
+        assert(self.nodes[0].getblockcount() == 201 + MAX_REORG_LENGTH)
 
         self.split_network()
 
         print("Test the minimum-rejected reorg:")
-        print("Mine 100 blocks on Node 0")
-        self.nodes[0].generate(100)
-        assert(self.nodes[0].getblockcount() == 400)
-        assert(self.nodes[2].getblockcount() == 300)
+        print("Mine %d blocks on Node 0" % (MAX_REORG_LENGTH + 1))
+        self.nodes[0].generate(MAX_REORG_LENGTH + 1)
+        assert(self.nodes[0].getblockcount() == 202 + 2 * MAX_REORG_LENGTH)
+        assert(self.nodes[2].getblockcount() == 201 + MAX_REORG_LENGTH)
 
-        print("Mine competing 101 blocks on Node 2")
-        self.nodes[2].generate(101)
-        assert(self.nodes[0].getblockcount() == 400)
-        assert(self.nodes[2].getblockcount() == 401)
+        print("Mine competing %d blocks on Node 2" % (MAX_REORG_LENGTH + 2))
+        self.nodes[2].generate(MAX_REORG_LENGTH + 2)
+        assert(self.nodes[0].getblockcount() == 202 + 2 * MAX_REORG_LENGTH)
+        assert(self.nodes[2].getblockcount() == 203 + 2 * MAX_REORG_LENGTH)
 
         try:
             print("Sync nodes to force a reorg")
@@ -95,7 +99,7 @@ class ReorgLimitTest(BitcoinTestFramework):
             # Check that node 0 stopped for the expected reason.
             self.log_stderr.seek(0)
             stderr = self.log_stderr.read().decode('utf-8')
-            expected_msg = "A block chain reorganization has been detected that would roll back 100 blocks!"
+            expected_msg = "A block chain reorganization has been detected that would roll back %d blocks!" % (MAX_REORG_LENGTH + 1)
             if expected_msg not in stderr:
                 raise AssertionError("Expected error \"" + expected_msg + "\" not found in:\n" + stderr)
         finally: